+++ apps/wallet/src/ui/app/hooks/deepbook/index.ts
@@ -1,0 +1,7 @@
+// Copyright (c) Mysten Labs, Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+export * from './useDeepBookConfigs';
+export * from './useRecognizedCoins';
+export * from './useDeepbookPools';
+export * from './useBalanceConversion';
