+++ apps/wallet/src/ui/app/hooks/index.ts
@@ -22,3 +22,4 @@ export * from './useTransactionData';
 export * from './useActiveAddress';
 export * from './useGetAllCoins';
 export * from './useCoinsReFetchingConfig';
+export * from './deepbook';
