+++ apps/wallet/src/ui/app/hooks/useDeepBook.ts
@@ -13,179 +13,13 @@ import {
 import { useDeepBookContext } from '_shared/deepBook/context';
 import { FEATURES } from '_shared/experimentation/features';
 import { useFeatureValue } from '@growthbook/growthbook-react';
-import { roundFloat, useGetObject } from '@mysten/core';
+import { useGetObject } from '@mysten/core';
 import { useSuiClient } from '@mysten/dapp-kit';
 import { type DeepBookClient } from '@mysten/deepbook';
 import { TransactionBlock } from '@mysten/sui.js/builder';
 import { type CoinStruct, type SuiClient } from '@mysten/sui.js/client';
-import { SUI_TYPE_ARG } from '@mysten/sui.js/utils';
 import { useQuery, useQueryClient } from '@tanstack/react-query';
 import BigNumber from 'bignumber.js';
-import { useMemo } from 'react';
-
-export enum Coins {
-	SUI = 'SUI',
-	USDC = 'USDC',
-	USDT = 'USDT',
-	WETH = 'WETH',
-	TBTC = 'TBTC',
-}
-
-export const mainnetDeepBook: {
-	pools: Record<string, string[]>;
-	coinsMap: Record<Coins, string>;
-} = {
-	pools: {
-		SUI_USDC: [
-			'0x7f526b1263c4b91b43c9e646419b5696f424de28dda3c1e6658cc0a54558baa7',
-			'0x18d871e3c3da99046dfc0d3de612c5d88859bc03b8f0568bd127d0e70dbc58be',
-		],
-		WETH_USDC: ['0xd9e45ab5440d61cc52e3b2bd915cdd643146f7593d587c715bc7bfa48311d826'],
-		TBTC_USDC: ['0xf0f663cf87f1eb124da2fc9be813e0ce262146f3df60bc2052d738eb41a25899'],
-		USDT_USDC: ['0x5deafda22b6b86127ea4299503362638bea0ca33bb212ea3a67b029356b8b955'],
-	},
-	coinsMap: {
-		[Coins.SUI]: '0x0000000000000000000000000000000000000000000000000000000000000002::sui::SUI',
-		[Coins.USDC]: '0x5d4b302506645c37ff133b98c4b50a5ae14841659738d6d733d59d0d217a93bf::coin::COIN',
-		[Coins.USDT]: '0xc060006111016b8a020ad5b33834984a437aaa7d3c74c18e09a95d48aceab08c::coin::COIN',
-		[Coins.WETH]: '0xaf8cd5edc19c4512f4259f0bee101a40d41ebed738ade5874359610ef8eeced5::coin::COIN',
-		[Coins.TBTC]: '0xbc3a676894871284b3ccfb2eec66f428612000e2a6e6d23f592ce8833c27c973::coin::COIN',
-	},
-};
-
-export function useDeepBookConfigs() {
-	return mainnetDeepBook;
-}
-
-export function useRecognizedCoins() {
-	const coinsMap = useDeepBookConfigs().coinsMap;
-	return Object.values(coinsMap);
-}
-
-export const allowedSwapCoinsList = [SUI_TYPE_ARG, mainnetDeepBook.coinsMap[Coins.USDC]];
-
-export function getUSDCurrency(amount: number | null) {
-	if (typeof amount !== 'number') {
-		return null;
-	}
-
-	return roundFloat(amount).toLocaleString('en', {
-		style: 'currency',
-		currency: 'USD',
-	});
-}
-
-export function useDeepbookPools() {
-	const deepBookClient = useDeepBookContext().client;
-
-	return useQuery({
-		queryKey: [DEEPBOOK_KEY, 'get-all-pools'],
-		queryFn: () => deepBookClient.getAllPools({}),
-	});
-}
-
-async function getDeepBookPriceForCoin(
-	coin: Coins,
-	pools: Record<string, string[]>,
-	isAsk: boolean,
-	deepBookClient: DeepBookClient,
-) {
-	if (coin === Coins.USDC) {
-		return 1n;
-	}
-
-	const poolName = `${coin}_USDC`;
-	const poolIds = pools[poolName];
-	const promises = poolIds.map(async (poolId) => {
-		const { bestBidPrice, bestAskPrice } = await deepBookClient.getMarketPrice(poolId);
-
-		return isAsk ? bestBidPrice : bestAskPrice;
-	});
-
-	const prices = await Promise.all(promises);
-
-	const filter: bigint[] = prices.filter((price): price is bigint => {
-		return typeof price === 'bigint' && price !== 0n;
-	});
-
-	const total = filter.reduce((acc, price) => {
-		return acc + price;
-	}, 0n);
-
-	return total / BigInt(filter.length);
-}
-
-function useDeepbookPricesInUSD(coins: Coins[], isAsk: boolean) {
-	const deepBookClient = useDeepBookContext().client;
-	const deepbookPools = useDeepBookConfigs().pools;
-
-	return useQuery({
-		queryKey: [DEEPBOOK_KEY, 'get-prices-usd', coins, isAsk],
-		queryFn: async () => {
-			const promises = coins.map((coin) =>
-				getDeepBookPriceForCoin(coin, deepbookPools, isAsk, deepBookClient),
-			);
-			return Promise.all(promises);
-		},
-	});
-}
-
-function useAveragePrice(base: Coins, quote: Coins, isAsk: boolean) {
-	const { data: prices, ...rest } = useDeepbookPricesInUSD([base, quote], isAsk);
-
-	const averagePrice = useMemo(() => {
-		const basePrice = new BigNumber((prices?.[0] ?? 1n).toString());
-		const quotePrice = new BigNumber((prices?.[1] ?? 1n).toString());
-
-		const basePriceBigNumber = new BigNumber(basePrice.toString());
-		const quotePriceBigNumber = new BigNumber(quotePrice.toString());
-
-		let avgPrice;
-		if (quote === Coins.USDC) {
-			avgPrice = basePriceBigNumber;
-		} else {
-			avgPrice = basePriceBigNumber.dividedBy(quotePriceBigNumber);
-		}
-
-		return avgPrice;
-	}, [prices, quote]);
-
-	return {
-		data: averagePrice,
-		...rest,
-	};
-}
-
-export function useBalanceConversion(
-	balance: BigInt | BigNumber | null,
-	from: Coins,
-	to: Coins,
-	conversionRate: number = 1,
-) {
-	const { data: averagePrice, ...rest } = useAveragePrice(from, to, to === Coins.USDC);
-
-	const averagePriceWithConversion = averagePrice.shiftedBy(conversionRate);
-
-	const rawValue = useMemo(() => {
-		if (!averagePriceWithConversion || !balance) return null;
-
-		const rawUsdValue = new BigNumber(balance.toString())
-			.multipliedBy(averagePriceWithConversion)
-			.toNumber();
-
-		if (isNaN(rawUsdValue)) {
-			return null;
-		}
-
-		return rawUsdValue;
-	}, [averagePriceWithConversion, balance]);
-
-	return {
-		rawValue,
-		averagePrice: averagePriceWithConversion,
-		...rest,
-	};
-}
 
 const MAX_COINS_PER_REQUEST = 10;
 
@@ -451,45 +285,4 @@ export function useGetEstimate({
 			!!signer &&
 			!!activeAddress,
 	});
-}
-
-export async function isExceedingSlippageTolerance({
-	slipPercentage,
-	poolId,
-	deepBookClient,
-	conversionRate,
-	baseCoinAmount,
-	quoteCoinAmount,
-	isAsk,
-}: {
-	slipPercentage: string;
-	poolId: string;
-	deepBookClient: DeepBookClient;
-	conversionRate: number;
-	baseCoinAmount?: string;
-	quoteCoinAmount?: string;
-	isAsk: boolean;
-}) {
-	if (!baseCoinAmount || !quoteCoinAmount) {
-		return false;
-	}
-
-	const bigNumberBaseCoinAmount = new BigNumber(baseCoinAmount).abs();
-	const bigNumberQuoteCoinAmount = new BigNumber(quoteCoinAmount).abs();
-
-	const averagePricePaid = bigNumberQuoteCoinAmount
-		.dividedBy(bigNumberBaseCoinAmount)
-		.shiftedBy(conversionRate);
-
-	const { bestBidPrice, bestAskPrice } = await deepBookClient.getMarketPrice(poolId);
-
-	if (!bestBidPrice || !bestAskPrice) {
-		return false;
-	}
-
-	const slip = new BigNumber(isAsk ? bestBidPrice.toString() : bestAskPrice.toString()).dividedBy(
-		averagePricePaid,
-	);
-
-	return new BigNumber('1').minus(slip).abs().isGreaterThan(slipPercentage);
 }
