+++ apps/wallet/src/ui/app/pages/home/tokens/TokensDetails.tsx
@@ -1,7 +1,7 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-import { allowedSwapCoinsList } from '_app/hooks/useDeepBook';
+import { allowedSwapCoinsList } from '_app/hooks/deepbook';
 import { useIsWalletDefiEnabled } from '_app/hooks/useIsWalletDefiEnabled';
 import { LargeButton } from '_app/shared/LargeButton';
 import { Text } from '_app/shared/text';
