+++ apps/wallet/src/ui/app/pages/swap/utils.ts
@@ -3,8 +3,9 @@
 import { useActiveAccount } from '_app/hooks/useActiveAccount';
 import { Coins, useBalanceConversion, useCoinsReFetchingConfig } from '_hooks';
 import { SUI_CONVERSION_RATE } from '_pages/swap/constants';
-import { useFormatCoin } from '@mysten/core';
+import { roundFloat, useFormatCoin } from '@mysten/core';
 import { useSuiClientQuery } from '@mysten/dapp-kit';
+import { type DeepBookClient } from '@mysten/deepbook';
 import BigNumber from 'bignumber.js';
 
 export function useSwapData({
@@ -76,4 +77,56 @@ export function useSuiUsdcBalanceConversion({ amount }: { amount: string }) {
 		suiUsdc,
 		usdcSui,
 	};
+}
+
+export function getUSDCurrency(amount: number | null) {
+	if (typeof amount !== 'number') {
+		return null;
+	}
+
+	return roundFloat(amount).toLocaleString('en', {
+		style: 'currency',
+		currency: 'USD',
+	});
+}
+
+export async function isExceedingSlippageTolerance({
+	slipPercentage,
+	poolId,
+	deepBookClient,
+	conversionRate,
+	baseCoinAmount,
+	quoteCoinAmount,
+	isAsk,
+}: {
+	slipPercentage: string;
+	poolId: string;
+	deepBookClient: DeepBookClient;
+	conversionRate: number;
+	baseCoinAmount?: string;
+	quoteCoinAmount?: string;
+	isAsk: boolean;
+}) {
+	if (!baseCoinAmount || !quoteCoinAmount) {
+		return false;
+	}
+
+	const bigNumberBaseCoinAmount = new BigNumber(baseCoinAmount).abs();
+	const bigNumberQuoteCoinAmount = new BigNumber(quoteCoinAmount).abs();
+
+	const averagePricePaid = bigNumberQuoteCoinAmount
+		.dividedBy(bigNumberBaseCoinAmount)
+		.shiftedBy(conversionRate);
+
+	const { bestBidPrice, bestAskPrice } = await deepBookClient.getMarketPrice(poolId);
+
+	if (!bestBidPrice || !bestAskPrice) {
+		return false;
+	}
+
+	const slip = new BigNumber(isAsk ? bestBidPrice.toString() : bestAskPrice.toString()).dividedBy(
+		averagePricePaid,
+	);
+
+	return new BigNumber('1').minus(slip).abs().isGreaterThan(slipPercentage);
 }
